add instructions for showing the sidebar (#70)
* add instructions for showing the sidebar

This will probably be obvious to someone who‘s used Nova for a while, but as a brand new user, it took me a while to find the TypeScript sidebar that’s mentioned in the docs. Based on the screenshot I decided it must be referring to the “Symbols” sidebar, but that wasn’t showing me any TypeScript features.

Suggestion is to add information to the documentation about where to find the sidebar. You may also want to mention the “Issues” sidebar.

* Update README.md

* More documentation about the sidebars

Co-authored-by: Cameron Little <cameron@camlittle.com>
Co-authored-by: Cameron Little <cameron@remitly.com>